Fix clang-tidy theme 1: narrowing conversions and implicit bool default
Add explicit static_cast<> for 5 narrowing-conversion diagnostics across
fftwguard.h (ptrdiff_t->int), radialbincomputer.h (double<->ptrdiff_t),
and randomgaussian.h (array init false positive). Change default argument
from true to 1 in radialprojectionsinglebinandvalue.h to fix implicit
bool->int conversion.

+++ src/include/TempLat/fft/external/fftw/fftwguard.h
@@ -49,9 +49,9 @@ namespace TempLat
       /* set the default: */
 
       if (mVerbose) sayShort << "Calling FFTW thread initializations.\n";
-      fftw_plan_with_nthreads(ThreadSettings::getMaxThreadCount());
+      fftw_plan_with_nthreads(static_cast<int>(ThreadSettings::getMaxThreadCount()));
 #ifdef HAVE_FFTFLOAT
-      fftwf_plan_with_nthreads(ThreadSettings::getMaxThreadCount());
+      fftwf_plan_with_nthreads(static_cast<int>(ThreadSettings::getMaxThreadCount()));
 #endif
 
 #ifdef HAVE_MPI
